Changes to Ice Servers and App.js
Using public Stun and Turn servers that I have tested to ensure that
they are still in service. Also, made change to the application to fix
the issue of calls being place prior to start button being clicked by
both clients. This commit should ensure that the app properly and fully
functions.